fabonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	int i, num1, num2, temp;
+	unsigned long i, num1, num2, temp;
 	
 	num1 = 0;
 	num2 = 1;
@@ -15,7 +15,7 @@ int main(void)
 		temp = num1 + num2;
 		num1 = num2;
 		num2 = temp;
-		printf("%d", temp);
+		printf("%lu", temp);
 		if (i == 49)
 			putchar('\n');
 		else
